Rename project to `fundom` (#85)
