+++ .github/workflows/python-package.yml
@@ -30,10 +30,10 @@ jobs:
           poetry install
       - name: Check flake8
         run: |
-          poetry run flake8 pymon --count --show-source --statistics
+          poetry run flake8 fundom --count --show-source --statistics
       - name: Check isort
         run: |
-          poetry run isort pymon/**/*.py --check-only --verbose
+          poetry run isort fundom/**/*.py --check-only --verbose
       - name: Test with pytest
         run: |
           poetry run pytest
