+++ docs/MODULES.md
@@ -2,16 +2,16 @@
 
 > Auto-generated documentation modules index.
 
-Full list of [Pymon](README.md#-pymon) project modules.
+Full list of [Pymon](README.md#-fundom) project modules.
 
-- [▶️ pymon](README.md#-pymon)
-- [Pymon](pymon/index.md#pymon)
-    - [Core](pymon/core.md#core)
-    - [Maybe](pymon/maybe.md#maybe)
-    - [Model](pymon/model.md#model)
-    - [Pointfree](pymon/pointfree/index.md#pointfree)
-        - [Bytes Utils](pymon/pointfree/bytes_utils.md#bytes-utils)
-        - [Dict Utils](pymon/pointfree/dict_utils.md#dict-utils)
-        - [Str Utils](pymon/pointfree/str_utils.md#str-utils)
-    - [Predicate](pymon/predicate.md#predicate)
-    - [Result](pymon/result.md#result)
+- [▶️ fundom](README.md#-fundom)
+- [Fundom](fundom/index.md#fundom)
+    - [Core](fundom/core.md#core)
+    - [Maybe](fundom/maybe.md#maybe)
+    - [Model](fundom/model.md#model)
+    - [Pointfree](fundom/pointfree/index.md#pointfree)
+        - [Bytes Utils](fundom/pointfree/bytes_utils.md#bytes-utils)
+        - [Dict Utils](fundom/pointfree/dict_utils.md#dict-utils)
+        - [Str Utils](fundom/pointfree/str_utils.md#str-utils)
+    - [Predicate](fundom/predicate.md#predicate)
+    - [Result](fundom/result.md#result)
