+++ docs/fundom/index.md
@@ -1,8 +1,8 @@
-# Pymon
+# Fundom
 
-> Auto-generated documentation for [pymon](https://github.com/katunilya/pymon/blob/main/pymon/__init__.py) module.
+> Auto-generated documentation for [fundom](https://github.com/katunilya/pymon/blob/main/fundom/__init__.py) module.
 
-- [Pymon](../README.md#-pymon) / [Modules](../MODULES.md#pymon-modules) / Pymon
+- [Pymon](../README.md#-fundom) / [Modules](../MODULES.md#pymon-modules) / Fundom
     - Modules
         - [Core](core.md#core)
         - [Maybe](maybe.md#maybe)
