+++ docs/fundom/pointfree/index.md
@@ -1,0 +1,9 @@
+# Pointfree
+
+> Auto-generated documentation for [fundom.pointfree](https://github.com/katunilya/pymon/blob/main/fundom/pointfree/__init__.py) module.
+
+- [Pymon](../../README.md#-fundom) / [Modules](../../MODULES.md#pymon-modules) / [Fundom](../index.md#fundom) / Pointfree
+    - Modules
+        - [Bytes Utils](bytes_utils.md#bytes-utils)
+        - [Dict Utils](dict_utils.md#dict-utils)
+        - [Str Utils](str_utils.md#str-utils)
