--- docs/pymon/pointfree/index.md
@@ -1,9 +1,0 @@
-# Pointfree
-
-> Auto-generated documentation for [pymon.pointfree](https://github.com/katunilya/pymon/blob/main/pymon/pointfree/__init__.py) module.
-
-- [Pymon](../../README.md#-pymon) / [Modules](../../MODULES.md#pymon-modules) / [Pymon](../index.md#pymon) / Pointfree
-    - Modules
-        - [Bytes Utils](bytes_utils.md#bytes-utils)
-        - [Dict Utils](dict_utils.md#dict-utils)
-        - [Str Utils](str_utils.md#str-utils)
