+++ fundom/pointfree/dict_utils.py
@@ -1,7 +1,7 @@
 from typing import TypeVar
 
-from pymon.core import hof1
-from pymon.result import safe
+from fundom.core import hof1
+from fundom.result import safe
 
 TKey = TypeVar("TKey")
 TValue = TypeVar("TValue")
